# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Palindot API",
        "description": "Api created for a URL Shortener that generates palindromic values.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Palindot API"
        }
    ],
    "paths": {
...
        "/urls/{url}": {
            "get": {
                "summary": "Gets a specific URL.",
                "description": "Display the specified resource.",
                "operationId": "31e04dec7e5c51a89fc4e7dc3d3d4150",
                "parameters": [
                    {
                        "name": "url",
                        "in": "path",
                        "description": "ID of the URL",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Specific URL"
                    }
                }
            },
            "put": {
                "summary": "Updates a specific URL.",
                "description": "Update the specified resource in storage.",
                "operationId": "971e217f911720a073ad8d3fd535dc7f",
                "parameters": [
                    {
                        "name": "url",
                        "in": "path",
                        "description": "ID of the URL",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "URL parameters",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UrlRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "URL updated successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "summary": "Deletes a specific URL.",
                "description": "Remove the specified resource from storage.",
                "operationId": "1e8da75470fb15db42fca61dec181437",
                "parameters": [
                    {
                        "name": "url",
                        "in": "path",
                        "description": "ID of the URL",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "URL deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/urls/create": {
            "post": {
                "summary": "Creates a new URL.",
                "description": "Store a newly created resource in storage.",
                "operationId": "8312d6ca405122041bba053d5f3da880",
                "requestBody": {
                    "description": "URL parameters",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UrlRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "URL created successfully"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "summary": "Displays the information of the current user.",
                "description": "Display the information of the current user.",
                "operationId": "7e2f5799553ec93c9c43e2a58490d447",
                "responses": {
                    "200": {
                        "description": "Authenticated user"
                    }
                }
            },
            "post": {
                "summary": "Creates a new user. Only available for the admin created via seeders.",
                "description": "Store a newly created resource in storage.",
                "operationId": "a1fce870c566eb5fc0cf3178ec5d2fed",
                "requestBody": {
                    "description": "User parameters",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "summary": "Displays the information of the current user, too lol.",
                "description": "Display the specified resource.",
                "operationId": "910a0c7cf580e1c605648d82a4e57f46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Specific user"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "summary": "Updates the current user.",
                "description": "Update the specified resource in storage.",
                "operationId": "ddfe232db1870b82493bf8bd25596d37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User parameters",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "summary": "Deletes the current user.",
                "description": "Remove the specified resource from storage.",
                "operationId": "43dd20becb1f47444e484cbb18ce345c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UrlRequest": {
                "description": "Requires the data necessary to store and update a url",
                "properties": {
                    "base_url": {
                        "description": "The url that you want to shorten",
                        "type": "string"
                    },
                    "subset_id": {
                        "description": "The subset in which you want to store the url. It can be null and it's not required",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserRequest": {
                "description": "Requires the data the admin needs to register an user",
                "properties": {
                    "name": {
                        "description": "The username",
                        "type": "string"
                    },
                    "email": {
                        "description": "The email of the new user",
                        "type": "string"
                    },
                    "password": {
                        "description": "The password of the new user",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
...
    }
}